# src/generated/openapi.generated.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "D&D Character Manager API",
		"version": "0.1.0"
	},
	"paths": {
...
		"/api/characters": {
			"post": {
				"operationId": "createCharacter",
				"summary": "Create character",
				"tags": [
					"characters"
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"minLength": 1,
										"maxLength": 120,
										"pattern": "\\S"
									},
									"className": {
										"type": "string",
										"enum": [
											"Barbarian",
											"Bard",
											"Cleric",
											"Druid",
											"Fighter",
											"Monk",
											"Paladin",
											"Ranger",
											"Rogue",
											"Sorcerer",
											"Warlock",
											"Wizard"
										]
									},
									"level": {
										"type": "integer",
										"minimum": 1,
										"maximum": 20
									},
									"maxHp": {
										"type": "integer",
										"minimum": 1,
										"maximum": 9999
									}
								},
								"required": [
									"name",
									"className",
									"level",
									"maxHp"
								],
								"additionalProperties": false
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Created character",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"character": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"format": "uuid",
													"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
												},
												"name": {
													"type": "string",
													"minLength": 1,
													"maxLength": 120,
													"pattern": "\\S"
												},
												"className": {
													"type": "string",
													"enum": [
														"Barbarian",
														"Bard",
														"Cleric",
														"Druid",
														"Fighter",
														"Monk",
														"Paladin",
														"Ranger",
														"Rogue",
														"Sorcerer",
														"Warlock",
														"Wizard"
													]
												},
												"level": {
													"type": "integer",
													"minimum": 1,
													"maximum": 20
												},
												"experiencePoints": {
													"type": "integer",
													"minimum": 0,
													"maximum": 9999999
												},
												"experience": {
													"type": "object",
													"properties": {
														"level": {
															"type": "integer",
															"minimum": 1,
															"maximum": 20
														},
														"experiencePoints": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999999
														},
														"currentLevelMinimum": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999999
														},
														"nextLevel": {
															"anyOf": [
																{
																	"type": "integer",
																	"minimum": 1,
																	"maximum": 20
																},
																{
																	"type": "null"
																}
															]
														},
														"nextLevelMinimum": {
															"anyOf": [
																{
																	"type": "integer",
																	"minimum": 0,
																	"maximum": 9999999
																},
																{
																	"type": "null"
																}
															]
														},
														"experienceIntoLevel": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999999
														},
														"experienceForNextLevel": {
															"anyOf": [
																{
																	"type": "integer",
																	"minimum": 0,
																	"maximum": 9999999
																},
																{
																	"type": "null"
																}
															]
														},
														"experienceRemaining": {
															"anyOf": [
																{
																	"type": "integer",
																	"minimum": 0,
																	"maximum": 9999999
																},
																{
																	"type": "null"
																}
															]
														},
														"progressPercent": {
															"type": "integer",
															"minimum": 0,
															"maximum": 100
														},
														"isMaxLevel": {
															"type": "boolean"
														}
													},
													"required": [
														"level",
														"experiencePoints",
														"currentLevelMinimum",
														"nextLevel",
														"nextLevelMinimum",
														"experienceIntoLevel",
														"experienceForNextLevel",
														"experienceRemaining",
														"progressPercent",
														"isMaxLevel"
													],
													"additionalProperties": false
												},
												"health": {
													"type": "object",
													"properties": {
														"currentHp": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999
														},
														"maxHp": {
															"type": "integer",
															"minimum": 1,
															"maximum": 9999
														},
														"temporaryHp": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999
														},
														"effectiveMaxHp": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999
														}
													},
													"required": [
														"currentHp",
														"maxHp",
														"temporaryHp",
														"effectiveMaxHp"
													],
													"additionalProperties": false
												},
												"recentHealthChanges": {
													"maxItems": 5,
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"format": "uuid",
																"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
															},
															"previous": {
																"type": "object",
																"properties": {
																	"currentHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	},
																	"maxHp": {
																		"type": "integer",
																		"minimum": 1,
																		"maximum": 9999
																	},
																	"temporaryHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	},
																	"effectiveMaxHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	}
																},
																"required": [
																	"currentHp",
																	"maxHp",
																	"temporaryHp",
																	"effectiveMaxHp"
																],
																"additionalProperties": false
															},
															"next": {
																"type": "object",
																"properties": {
																	"currentHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	},
																	"maxHp": {
																		"type": "integer",
																		"minimum": 1,
																		"maximum": 9999
																	},
																	"temporaryHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	},
																	"effectiveMaxHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	}
																},
																"required": [
																	"currentHp",
																	"maxHp",
																	"temporaryHp",
																	"effectiveMaxHp"
																],
																"additionalProperties": false
															},
															"currentHpDelta": {
																"type": "integer",
																"minimum": -9007199254740991,
																"maximum": 9007199254740991
															},
															"maxHpDelta": {
																"type": "integer",
																"minimum": -9007199254740991,
																"maximum": 9007199254740991
															},
															"temporaryHpDelta": {
																"type": "integer",
																"minimum": -9007199254740991,
																"maximum": 9007199254740991
															},
															"createdAt": {
																"type": "string",
																"format": "date-time",
																"pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
															}
														},
														"required": [
															"id",
															"previous",
															"next",
															"currentHpDelta",
															"maxHpDelta",
															"temporaryHpDelta",
															"createdAt"
														],
														"additionalProperties": false
													}
												}
											},
											"required": [
												"id",
												"name",
												"className",
												"level",
												"experiencePoints",
												"experience",
												"health",
												"recentHealthChanges"
											],
											"additionalProperties": false
										}
									},
									"required": [
										"character"
									],
									"additionalProperties": false
								}
							}
						}
					},
					"400": {
						"description": "Invalid character data",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"additionalProperties": false
								}
							}
						}
					}
				}
			},
			"get": {
				"operationId": "listCharacters",
				"summary": "List characters",
				"tags": [
					"characters"
				],
				"responses": {
					"200": {
						"description": "Characters",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"characters": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": {
														"type": "string",
														"format": "uuid",
														"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
													},
													"name": {
														"type": "string",
														"minLength": 1,
														"maxLength": 120,
														"pattern": "\\S"
													},
													"className": {
														"type": "string",
														"enum": [
															"Barbarian",
															"Bard",
															"Cleric",
															"Druid",
															"Fighter",
															"Monk",
															"Paladin",
															"Ranger",
															"Rogue",
															"Sorcerer",
															"Warlock",
															"Wizard"
														]
													},
													"level": {
														"type": "integer",
														"minimum": 1,
														"maximum": 20
													}
												},
												"required": [
													"id",
													"name",
													"className",
													"level"
												],
												"additionalProperties": false
											}
										}
									},
									"required": [
										"characters"
									],
									"additionalProperties": false
								}
							}
						}
					}
				}
			}
		},
		"/api/characters/{characterId}": {
			"get": {
				"operationId": "getCharacter",
				"summary": "Get character",
				"tags": [
					"characters"
				],
				"parameters": [
					{
						"name": "characterId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid",
							"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Character detail",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"character": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"format": "uuid",
													"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
												},
												"name": {
													"type": "string",
													"minLength": 1,
													"maxLength": 120,
													"pattern": "\\S"
												},
												"className": {
													"type": "string",
													"enum": [
														"Barbarian",
														"Bard",
														"Cleric",
														"Druid",
														"Fighter",
														"Monk",
														"Paladin",
														"Ranger",
														"Rogue",
														"Sorcerer",
														"Warlock",
														"Wizard"
													]
												},
												"level": {
													"type": "integer",
													"minimum": 1,
													"maximum": 20
												},
												"experiencePoints": {
													"type": "integer",
													"minimum": 0,
													"maximum": 9999999
												},
												"experience": {
													"type": "object",
													"properties": {
														"level": {
															"type": "integer",
															"minimum": 1,
															"maximum": 20
														},
														"experiencePoints": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999999
														},
														"currentLevelMinimum": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999999
														},
														"nextLevel": {
															"anyOf": [
																{
																	"type": "integer",
																	"minimum": 1,
																	"maximum": 20
																},
																{
																	"type": "null"
																}
															]
														},
														"nextLevelMinimum": {
															"anyOf": [
																{
																	"type": "integer",
																	"minimum": 0,
																	"maximum": 9999999
																},
																{
																	"type": "null"
																}
															]
														},
														"experienceIntoLevel": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999999
														},
														"experienceForNextLevel": {
															"anyOf": [
																{
																	"type": "integer",
																	"minimum": 0,
																	"maximum": 9999999
																},
																{
																	"type": "null"
																}
															]
														},
														"experienceRemaining": {
															"anyOf": [
																{
																	"type": "integer",
																	"minimum": 0,
																	"maximum": 9999999
																},
																{
																	"type": "null"
																}
															]
														},
														"progressPercent": {
															"type": "integer",
															"minimum": 0,
															"maximum": 100
														},
														"isMaxLevel": {
															"type": "boolean"
														}
													},
													"required": [
														"level",
														"experiencePoints",
														"currentLevelMinimum",
														"nextLevel",
														"nextLevelMinimum",
														"experienceIntoLevel",
														"experienceForNextLevel",
														"experienceRemaining",
														"progressPercent",
														"isMaxLevel"
													],
													"additionalProperties": false
												},
												"health": {
													"type": "object",
													"properties": {
														"currentHp": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999
														},
														"maxHp": {
															"type": "integer",
															"minimum": 1,
															"maximum": 9999
														},
														"temporaryHp": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999
														},
														"effectiveMaxHp": {
															"type": "integer",
															"minimum": 0,
															"maximum": 9999
														}
													},
													"required": [
														"currentHp",
														"maxHp",
														"temporaryHp",
														"effectiveMaxHp"
													],
													"additionalProperties": false
												},
												"recentHealthChanges": {
													"maxItems": 5,
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"format": "uuid",
																"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
															},
															"previous": {
																"type": "object",
																"properties": {
																	"currentHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	},
																	"maxHp": {
																		"type": "integer",
																		"minimum": 1,
																		"maximum": 9999
																	},
																	"temporaryHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	},
																	"effectiveMaxHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	}
																},
																"required": [
																	"currentHp",
																	"maxHp",
																	"temporaryHp",
																	"effectiveMaxHp"
																],
																"additionalProperties": false
															},
															"next": {
																"type": "object",
																"properties": {
																	"currentHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	},
																	"maxHp": {
																		"type": "integer",
																		"minimum": 1,
																		"maximum": 9999
																	},
																	"temporaryHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	},
																	"effectiveMaxHp": {
																		"type": "integer",
																		"minimum": 0,
																		"maximum": 9999
																	}
																},
																"required": [
																	"currentHp",
																	"maxHp",
																	"temporaryHp",
																	"effectiveMaxHp"
																],
																"additionalProperties": false
															},
															"currentHpDelta": {
																"type": "integer",
																"minimum": -9007199254740991,
																"maximum": 9007199254740991
															},
															"maxHpDelta": {
																"type": "integer",
																"minimum": -9007199254740991,
																"maximum": 9007199254740991
															},
															"temporaryHpDelta": {
																"type": "integer",
																"minimum": -9007199254740991,
																"maximum": 9007199254740991
															},
															"createdAt": {
																"type": "string",
																"format": "date-time",
																"pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
															}
														},
														"required": [
															"id",
															"previous",
															"next",
															"currentHpDelta",
															"maxHpDelta",
															"temporaryHpDelta",
															"createdAt"
														],
														"additionalProperties": false
													}
												}
											},
											"required": [
												"id",
												"name",
												"className",
												"level",
												"experiencePoints",
												"experience",
												"health",
												"recentHealthChanges"
											],
											"additionalProperties": false
										}
									},
									"required": [
										"character"
									],
									"additionalProperties": false
								}
							}
						}
					},
					"404": {
						"description": "Character not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"additionalProperties": false
								}
							}
						}
					}
				}
			}
		},
...
	}
}